feat : 카테고리, 정렬, 검색 각각 클릭시 url에 보여주기
