+++ src/components/home/Category/CategoryList.tsx
@@ -4,14 +4,10 @@ import { useQuery } from "@tanstack/react-query";
 import { getCategoryList } from "@/src/apis/category";
 import Category from "./category";
 import * as S from "./Styled/StyledCategoryList";
-
-export type filterSearchProps = {
-  category?: string;
-  sort?: string;
-  searchQuery?: string;
-};
+import useFilterSearch, { filterSearchProps } from "../../search/useFilterSearch";
 
 export default function CategoryList() {
+  const filterSearch = useFilterSearch();
   const { data: categoryList } = useQuery({
     queryKey: ["categoryList"],
     queryFn: () => getCategoryList(),
@@ -21,20 +17,7 @@ export default function CategoryList() {
   }
 
   const onChange = (e) => {
-    console.log(e.target.value);
     filterSearch({ category: e.target.value });
-  };
-  const router = useRouter();
-  const filterSearch = ({ category, sort, searchQuery }: filterSearchProps) => {
-    const { query } = router;
-    if (searchQuery !== undefined) query.searchQuery = searchQuery;
-    if (sort !== undefined) query.sortValue = sort;
-    if (category !== undefined) query.category = category;
-
-    router.push({
-      pathname: router.pathname,
-      query: query,
-    });
   };
 
   return (
